requirements.txt
tensorflow==2.11.0
keras==2.11.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.3
numpy==1.23.5
scikit-learn==1.2.2
joblib==1.2.0
pyjwt==2.8.0
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
streamlit==1.47.0
protobuf==3.20.0
insightface==0.7.3
onnxruntime==1.16.3
